Split city.py to different files & also move unused files to a different folder
