+++ src/citysimulation/__init__.py
@@ -1,1 +1,1 @@
-from .window import create_window
+from ..unused_prototypes.window import create_window
