+++ src/citysimulation/fractal_coords.py
@@ -1,0 +1,72 @@
+""" road_coordinates = [
+    # Horizontal Roads
+    ((50, 100), (250, 100)),
+    ((300, 200), (600, 200)),
+    ((100, 300), (400, 300)),
+    ((450, 400), (750, 400)),
+    ((200, 500), (500, 500)),
+
+    # Vertical Roads
+    ((100, 50), (100, 250)),
+    ((250, 150), (250, 450)),
+    ((400, 100), (400, 350)),
+    ((550, 200), (550, 500)),
+    ((700, 50), (700, 250)),
+
+    # Diagonal Roads
+    ((50, 50), (200, 200)),
+    ((250, 100), (400, 250)),
+    ((150, 400), (300, 550)),
+    ((500, 50), (650, 200)),
+    ((350, 300), (500, 450)),
+    ((600, 400), (750, 550)),
+    ((100, 600), (250, 750)),
+    ((400, 500), (550, 650)),
+    ((200, 250), (350, 400)),
+    ((650, 150), (800, 300))
+] """
+
+coordinates = [
+    (50, 100),
+    (250, 100),
+    (300, 200),
+    (600, 200),
+    (100, 300),
+    (400, 300),
+    (450, 400),
+    (750, 400),
+    (200, 500),
+    (500, 500),
+
+    (100, 50),
+    (100, 250),
+    (250, 150),
+    (250, 450),
+    (400, 100),
+    (400, 350),
+    (550, 200),
+    (550, 500),
+    (700, 50),
+    (700, 250),
+
+    (50, 50),
+    (200, 200),
+    (250, 100),
+    (400, 250),
+    (150, 400),
+    (300, 550),
+    (500, 50),
+    (650, 200),
+    (350, 300),
+    (500, 450),
+    (600, 400),
+    (750, 550),
+    (100, 600),
+    (250, 750),
+    (400, 500),
+    (550, 650),
+    (200, 250),
+    (350, 400),
+    (650, 150),
+    (800, 300)
+]
